Fix a catastrophic bug in the handling of return type polymorphism
The problem is: a function is said to be polymorphic on the return type iff there are type parameters in the return type that are not present in the list of *value* parameters. Thus: we can't figure out the return type by propagating from the types of the arguments, and we need contextual information (essentially, the programmer has to assert the return type).

The code works like this: take the set of type parameters in the value parameter list, P, take the set of type parameters in the return type, R, then do R-P. If that's anything other than the empty set: there are parameters in the return type that are not in the value parameter list, therefore, the function is return type polymorphic.

The problem is the code was taking the set of type parameters from the *type parameter list* rather than the *value parameter list*. This commit fixes this.