Ensure import for ToFloat and Remainder.
Discussion in Discord explored that `toFloat64(Int32)` did not appear to
be working properly. Found that although `ToFloat32`, `ToFloat64`, and
`Remainder` were implemented in Pervasive, they were not recognized in
applications. Verified the rest of the compiler pipeline and also
verified that a test application using `toFloat64(Int32)` worked
properly after applying this change.