Emacs mode: custom syntax-table to avoid interference.
font-lock first runs a pass using the character-by-character types
defined in syntax-table, which includes marking the double quote as a
string start character.
The regexps in font-lock-defaults, which include the comments regexp,
run later so if there is a quoted string inside the comment, the whole
line will be marked as a normal line with a string literal in it.

Also, the default syntax-table marks _ and & as “symbol”, not “word”.
That makes dabbrev-expand for instance to fail if you type first
&variable and then type va and request expansion: Emacs will reply that
there is no completion for it, because it considers the first appearance
of the symbol to start with & which is not present in the partial
instance to be completed.

The _ causes moving word-by-word, for instance with CTRL+Right, to stop
at the _ as if it was punctuation, instead of jumping over the whole
name. For instance with mut_root, starting at the left and pressing
CTRL+Right stops after mut, instead of jumping all the way to the end
after mut_root.

To avoid that, this defines _ as “word” (subset of “symbol”), & as
“punctuation”, etc.