Ensure import for ToFloat and Remainder. (#575)
* Ensure import for ToFloat and Remainder.

Discussion in Discord explored that `toFloat64(Int32)` did not appear to
be working properly. Found that although `ToFloat32`, `ToFloat64`, and
`Remainder` were implemented in Pervasive, they were not recognized in
applications. Verified the rest of the compiler pipeline and also
verified that a test application using `toFloat64(Int32)` worked
properly after applying this change.

* Add trivial tests for ToFloat32 and ToFloat64 import. (#1)

---------

Co-authored-by: pfm <pfm@garrity.co>
Co-authored-by: Alberto González Palomo <bugs@sentido-labs.com>